fix: no more blurry modal on chrome
The modal is now centered using flexbox instead of the translate -50% hack... when flexbox is available.
I added a custom copy of modernizr to do so, but it doesn't overwrite any modernizr you might have.
I also fixed a bug that would prevent click on backdrop to close the modal if you didn't have jquery included.